Add files via upload
PowerShell script that creates a scheduled task that creates/runs a PowerShell script that restarts the NIC adapter titled "Ethernet" on startup. This script was created to workaround a known issue with Windows Server 2025 and Domain Controllers where their network profile is stuck as Public until the network adapters have been restarted.